Issue #2 closed: URL typo and additional images
Dawnmarie, I addressed the fixes for the URL typo on Exercise 3 Task 3, and the request for additional screenshots on Exercise 3 Task 4.  I made these changes on my repo tied to Git Desktop before I realized that this was on the main and not the branch. I didn't want to lose my work, so I am commiting to the main.  I will get this reconfigured for the next time.  Thank you.  Dwayne

Co-Authored-By: Dawnmarie DesJardins <39317565+dawnmariedesjardins@users.noreply.github.com>